VergiLens/services/rag-brain/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
# RAG ve LLM Araçları
llama-index-core
llama-index-readers-file
llama-index-embeddings-huggingface
llama-index-llms-ollama
qdrant-client
sentence-transformers